// src/packages/types/report/SerReportReceivableDebitRpt.ts
export interface SerReportReceivableDebitRpt {
  TenKhachHang: string;
  DiaChi: string;
  DienThoai: string;
  GhiChu: string;
  TongSoTien: string;
  Lst_Ser_ReportReceivableDebitRpt: any;
}

export interface Search_SerReportReceivableDebitRpt_Params {
  DealerCode: any;
  FlagDataWH: any;
  ToDate: any;
}